refactor(api-server): tRPC-over-WebSocket transport + one composition root
Adds a WebSocket transport for the tRPC surface as a 1:1 alternative to the
HTTP one, and the api-server structural cleanup it motivated. No behavior
change to existing surfaces. Foundation for #3170 (live UI updates) — the
subscription procedures + event fan-out land in follow-ups on top.

Transport:
- tRPC-over-WebSocket endpoint at /api/trpc-ws: same appRouter, wsLink
  transport, first-frame connectionParams auth (no ?token= in the URL),
  credential-bounded connections (reconnect nudge before token exp, hard
  close at it; API-key connections re-auth on a fixed 5min cadence), graceful
  drain on shutdown. Queries/mutations work over it today; subscriptions
  become its first new consumer later.

Structure:
- apps/api-server split into admission/ (auth + terms gates, shared
  authentication leg, per-edge denial mappers), trpc/ (http + ws doors,
  context factory), agent-proxies/ (relays, URLPattern upgrade dispatch,
  trpc/import proxies), routes/ (bootstrap endpoints, brand); module-owned
  routes (oauth, slack, telegram, artifacts) moved to their modules with
  narrowed prefixes.
- one composition root: bootstrap.ts assembles every singleton and returns
  per-app deps + a cleanup closure; index.ts spawns the three servers and
  conducts shutdown; app.ts is pure surface wiring. Fixes a latent divergence
  (a duplicated connectionsBoot was missing operatorCredentials on one side).
- shared authentication leg threaded to every edge; UserAuthenticated emitted
  per WS connection too.
- removed the shipped l7-promotion backfill migration.
