+++ packages/api-server/src/apps/api-server/admission/auth.ts
@@ -1,15 +1,16 @@
+import type { IncomingMessage } from "node:http";
 import { createRemoteJWKSet, jwtVerify, type JWTPayload } from "jose";
 import { JOSEError, JWKSInvalid, JWKSTimeout } from "jose/errors";
-import type { Context, MiddlewareHandler } from "hono";
-import { ALL_SCOPES, type UserIdentity } from "api-server-api";
-import type { Result } from "../../core/result.js";
-import { emit, EventType } from "../../events.js";
-import { securityLog } from "../../core/security-log.js";
+import type { Context } from "hono";
+import { ALL_SCOPES, type Scope, type UserIdentity } from "api-server-api";
+import type { Result } from "../../../core/result.js";
+import { securityLog } from "../../../core/security-log.js";
+import { emit, EventType } from "../../../events.js";
 import {
   isApiKeyToken,
   type ApiKeyValidationFailure,
   type ValidatedApiKey,
-} from "../../modules/api-keys/index.js";
+} from "../../../modules/api-keys/index.js";
 
 export class ForbiddenError extends Error {
   constructor(
@@ -27,6 +28,17 @@ export function clientIp(c: Context): string | undefined {
   const fwd = c.req.header("x-forwarded-for");
   if (fwd) return fwd.split(",")[0]!.trim();
   return c.req.header("x-real-ip") ?? undefined;
+}
+
+/** First-mile client IP of an upgrade request (the `clientIp` twin for
+ *  handlers that see the raw request instead of a Hono context). */
+export function upgradeSourceIp(req: IncomingMessage): string | undefined {
+  const fwd = req.headers["x-forwarded-for"];
+  return (
+    (typeof fwd === "string" ? fwd.split(",")[0]!.trim() : undefined) ??
+    req.socket.remoteAddress ??
+    undefined
+  );
 }
 
 export class UnauthorizedError extends Error {
@@ -120,25 +132,19 @@ export interface AuthDeps {
   verifyOwnerActive?: (sub: string) => Promise<boolean>;
 }
 
-const PUBLIC_PATHS = new Set([
-  "/api/health",
-  "/api/ready",
-  "/api/auth/config",
-  "/api/oauth/callback",
-  "/api/slack/oauth/callback",
-  "/api/telegram/oauth/callback",
-  "/api/terms",
-]);
-
-const PUBLIC_PATH_PREFIXES = ["/api/brand/"];
-
 /** Resolved principal plus the JWT-only telemetry signals (`azp`, realm
  *  `roles`). API-key principals carry empty values for both — they are not
  *  JWTs, so they have no `azp` and no realm roles. */
-interface VerifiedPrincipal {
+export interface VerifiedPrincipal {
   user: UserIdentity;
   azp: string;
   roles: string[];
+  /** When the verified credential stops being valid. Consumers that hold a
+   *  connection open past a single request (WS relays, the events socket)
+   *  must not let it outlive this instant. Absent for API keys — their
+   *  expiry and revocation are enforced per verify call, so connection
+   *  owners apply their own re-auth deadline instead. */
+  expiresAt?: Date;
 }
 
 export function createAuth(config: AuthConfig, deps: AuthDeps = {}) {
@@ -177,6 +183,8 @@ export function createAuth(config: AuthConfig, deps: AuthDeps = {}) {
       },
       azp: typeof claims.azp === "string" ? claims.azp : "",
       roles,
+      expiresAt:
+        payload.exp !== undefined ? new Date(payload.exp * 1000) : undefined,
     };
   }
 
@@ -216,107 +224,254 @@ export function createAuth(config: AuthConfig, deps: AuthDeps = {}) {
     return isApiKeyToken(token) ? verifyApiKey(token) : verifyJwt(token);
   }
 
-  const middleware: MiddlewareHandler = async (c, next) => {
-    if (
-      PUBLIC_PATHS.has(c.req.path) ||
-      PUBLIC_PATH_PREFIXES.some((p) => c.req.path.startsWith(p))
-    )
-      return next();
+  // `reload()` fetches the JWKS immediately and populates jose's cache —
+  // the boot-time warm hook behind `/api/ready` (see jwks-warmup.ts).
+  return { verify, warmJwks: () => JWKS.reload() };
+}
 
-    const authHeader = c.req.header("authorization");
-    if (!authHeader?.startsWith("Bearer ")) {
+// ---------------------------------------------------------------------------
+// Principal-authorization helpers (non-tRPC surfaces)
+// ---------------------------------------------------------------------------
+
+/** Binding check for non-tRPC surfaces (in-pod relay, WS upgrade,
+ *  import proxy). Returns true when the principal may operate `agentId`. */
+export function hasAgentBinding(user: UserIdentity, agentId: string): boolean {
+  return user.agentIds === "*" || user.agentIds.includes(agentId);
+}
+
+/** Scope guard for non-tRPC surfaces. tRPC routers use the
+ *  procedure builders in api-server-api/auth-procedures.ts. */
+export function hasScope(user: UserIdentity, scope: Scope): boolean {
+  return user.scopes.includes(scope);
+}
+
+// ---------------------------------------------------------------------------
+// The shared authentication leg
+// ---------------------------------------------------------------------------
+// Every edge admits a principal the same way; what differs is where the
+// token rides (HTTP header, query string, first WS frame) and how a denial
+// is delivered. Classification and the audit vocabulary live here, once;
+// each edge encodes the returned kind through its folder's mappers.
+
+/** Where an authentication attempt happened. Drives which audit vocabulary
+ *  the log line uses — event names and field shapes are a contract for
+ *  audit-trail consumers, so each edge keeps its established lines. */
+export type AuthSite =
+  | { edge: "http"; target: string; sourceIp?: string }
+  | {
+      edge: "ws";
+      /** Which WS surface: "acp" | "terminal" | "ssh" | "trpc". */
+      relay: string;
+      agentId?: string;
+      sourceIp?: string;
+    };
+
+/** The WS flavor of AuthSite — socket surfaces build it once per connection
+ *  and reuse it for every gate and the attach log. */
+export type WsAuthSite = Extract<AuthSite, { edge: "ws" }>;
+
+/** Why authentication refused a principal — semantic only, and only the
+ *  kinds this file can produce (other gates own theirs, e.g. terms'
+ *  "terms-stale"). Wire encoding belongs to the edge that delivers the
+ *  denial, through its folder's mappers. */
+export type AuthDenialKind =
+  | "missing-token"
+  | "auth-unavailable"
+  | "unauthorized"
+  | "forbidden";
+
+export type AuthenticateResult =
+  | { ok: true; principal: VerifiedPrincipal }
+  | { ok: false; kind: AuthDenialKind };
+
+/** The one authentication decision, bound to a verifier — defined once in
+ *  app.ts and threaded into every edge (HTTP middleware, relay admission,
+ *  tRPC WS endpoint), so no edge can construct its own. */
+export type Authenticate = (
+  token: string | null | undefined,
+  site: AuthSite,
+) => Promise<AuthenticateResult>;
+
+/** OIDC client ids matched against JWT `azp` to attribute an authentication
+ *  to surface="ui" / "cli"; `coreRole` flags core-team users. */
+export interface SurfaceAttribution {
+  uiClientId: string;
+  cliClientId: string;
+  coreRole?: string;
+}
+
+/** The usage-analytics fact "a credential was verified for this principal" —
+ *  emitted once per HTTP request (auth middleware) and once per WS
+ *  connection (tRPC WS admission). Feeds the auth Activity Event and
+ *  `actor_roles.is_core`; one helper so the emission sites cannot drift. */
+export function emitUserAuthenticated(
+  principal: VerifiedPrincipal,
+  attribution: SurfaceAttribution,
+): void {
+  const surface =
+    principal.azp === attribution.uiClientId
+      ? "ui"
+      : principal.azp === attribution.cliClientId
+        ? "cli"
+        : "other";
+  const isCore = attribution.coreRole
+    ? principal.roles.includes(attribution.coreRole)
+    : false;
+  emit({
+    type: EventType.UserAuthenticated,
+    userSub: principal.user.sub,
+    surface,
+    isCore,
+  });
+}
+
+function wsLogBase(site: WsAuthSite) {
+  return {
+    surface: "ws" as const,
+    ...(site.agentId !== undefined ? { agentId: site.agentId } : {}),
+    sourceIp: site.sourceIp,
+    detail: { relay: site.relay },
+  };
+}
+
+/** The authentication leg every edge shares — token presence and `verify` —
+ *  with classification and the deny audit vocabulary in one place so edges
+ *  cannot drift. Strictly "who are you": callers compose their own gates
+ *  after it (terms, owner, scope) and deliver every denial through their
+ *  own mapper. */
+export async function authenticatePrincipal(
+  verify: (token: string) => Promise<VerifiedPrincipal>,
+  token: string | null | undefined,
+  site: AuthSite,
+): Promise<AuthenticateResult> {
+  if (!token) {
+    if (site.edge === "http") {
       securityLog("warn", "authn.deny", {
         category: "authn",
         actor: null,
         actorKind: "external",
         result: "failure",
         reason: "missing-bearer",
-        target: c.req.path,
-        sourceIp: clientIp(c),
+        target: site.target,
+        sourceIp: site.sourceIp,
       });
-      return c.json({ error: "unauthorized" }, 401);
+    } else {
+      securityLog("warn", "ws.authn_deny", {
+        category: "authn",
+        actor: null,
+        actorKind: "external",
+        decision: "deny",
+        reason: "missing-token",
+        ...wsLogBase(site),
+      });
     }
+    return { ok: false, kind: "missing-token" };
+  }
 
-    const token = authHeader.slice(7);
-    try {
-      const { user, azp, roles } = await verify(token);
-      c.set("user", user);
-      c.set("roles", roles);
-      const surface =
-        azp === config.uiClientId
-          ? "ui"
-          : azp === config.cliClientId
-            ? "cli"
-            : "other";
-      const isCore = config.coreRole ? roles.includes(config.coreRole) : false;
-      emit({
-        type: EventType.UserAuthenticated,
-        userSub: user.sub,
-        surface,
-        isCore,
-      });
-      return next();
-    } catch (err) {
-      if (err instanceof AuthUnavailableError) {
-        // No credential verdict was reached (hence no `decision` field): the
-        // JWKS fetch failed, so signal a retryable outage, not a rejection.
+  let principal: VerifiedPrincipal;
+  try {
+    principal = await verify(token);
+  } catch (err) {
+    if (err instanceof AuthUnavailableError) {
+      // No credential verdict was reached (hence no `decision` field): the
+      // JWKS fetch failed, so signal a retryable outage, not a rejection.
+      if (site.edge === "http") {
         securityLog("warn", "authn.unavailable", {
           category: "authn",
           actor: null,
           actorKind: "external",
           result: "failure",
           reason: err.reason,
-          target: c.req.path,
-          sourceIp: clientIp(c),
+          target: site.target,
+          sourceIp: site.sourceIp,
           detail: { cause: describeJwksFailure(err.cause) },
         });
-        return c.json({ error: "auth unavailable" }, 503);
+      } else {
+        securityLog("warn", "ws.authn_unavailable", {
+          category: "authn",
+          actor: null,
+          actorKind: "external",
+          result: "failure",
+          reason: err.reason,
+          ...wsLogBase(site),
+        });
       }
-      if (err instanceof ForbiddenError) {
-        // Known principal denied for lack of a required role — the most
-        // forensically interesting authz event.
+      return { ok: false, kind: "auth-unavailable" };
+    }
+    if (err instanceof ForbiddenError) {
+      // Known principal denied for lack of a required role — the most
+      // forensically interesting authz event.
+      if (site.edge === "http") {
         securityLog("warn", "authz.deny", {
           category: "authz",
           actor: err.sub,
           actorKind: "user",
           result: "failure",
           reason: "missing-required-role",
-          target: c.req.path,
-          sourceIp: clientIp(c),
+          target: site.target,
+          sourceIp: site.sourceIp,
           detail: { requiredRole: err.requiredRole },
         });
-        return c.json(
-          {
-            error: "forbidden",
-            message: "Access pending approval. Contact your administrator.",
-          },
-          403,
-        );
+      } else {
+        securityLog("warn", "ws.authz_deny", {
+          category: "authz",
+          actor: err.sub,
+          actorKind: "user",
+          decision: "deny",
+          reason: "missing-required-role",
+          ...wsLogBase(site),
+        });
       }
-      // Token present but invalid — log the failure class (never the token
-      // itself): expired/bad-signature/wrong-audience are replay and tampering
-      // signals; API-key denials carry their reason string ("revoked", "owner
-      // inactive", …) so key misuse is distinguishable from JWT failures in
-      // the audit trail.
+      return { ok: false, kind: "forbidden" };
+    }
+    // Token present but invalid — log the failure class (never the token
+    // itself): expired/bad-signature/wrong-audience are replay and tampering
+    // signals; API-key denials carry their reason string ("revoked", "owner
+    // inactive", …) so key misuse is distinguishable from JWT failures in
+    // the audit trail.
+    const reason =
+      err instanceof UnauthorizedError
+        ? err.reason
+        : err instanceof Error
+          ? err.name
+          : "verify-failed";
+    if (site.edge === "http") {
       securityLog("warn", "authn.deny", {
         category: "authn",
         actor: null,
         actorKind: "external",
         result: "failure",
-        reason:
-          err instanceof UnauthorizedError
-            ? err.reason
-            : err instanceof Error
-              ? err.name
-              : "verify-failed",
-        target: c.req.path,
-        sourceIp: clientIp(c),
+        reason,
+        target: site.target,
+        sourceIp: site.sourceIp,
       });
-      return c.json({ error: "unauthorized" }, 401);
+    } else {
+      securityLog("warn", "ws.authn_deny", {
+        category: "authn",
+        actor: null,
+        actorKind: "external",
+        decision: "deny",
+        reason,
+        ...wsLogBase(site),
+      });
     }
-  };
+    return { ok: false, kind: "unauthorized" };
+  }
 
-  // `reload()` fetches the JWKS immediately and populates jose's cache —
-  // the boot-time warm hook behind `/api/ready` (see jwks-warmup.ts).
-  return { middleware, verify, warmJwks: () => JWKS.reload() };
+  return { ok: true, principal };
+}
+
+/** The high-value forensic event every WS surface emits once fully
+ *  admitted: someone is attached and live. */
+export function logWsAttach(sub: string, site: WsAuthSite): void {
+  securityLog("info", "relay.attach", {
+    category: "privileged",
+    actor: sub,
+    actorKind: "user",
+    surface: "ws",
+    ...(site.agentId !== undefined ? { agentId: site.agentId } : {}),
+    result: "success",
+    sourceIp: site.sourceIp,
+    detail: { relay: site.relay },
+  });
 }
