+++ packages/api-server/src/apps/api-server/agent-proxies/ssh-relay.ts
@@ -1,10 +1,10 @@
 import { WebSocketServer, WebSocket } from "ws";
 import type { IncomingMessage } from "node:http";
 import type { Duplex } from "node:stream";
-import { podBaseUrl } from "../../modules/agents/infrastructure/k8s.js";
-import type { AgentsRepository } from "../../modules/agents/infrastructure/agents-repository.js";
-import { isAgentWakeTimeoutError } from "../../modules/agents/index.js";
-import { LAST_ACTIVITY_KEY } from "../../modules/agents/infrastructure/labels.js";
+import { podBaseUrl } from "../../../modules/agents/infrastructure/k8s.js";
+import type { AgentsRepository } from "../../../modules/agents/infrastructure/agents-repository.js";
+import { isAgentWakeTimeoutError } from "../../../modules/agents/index.js";
+import { LAST_ACTIVITY_KEY } from "../../../modules/agents/infrastructure/labels.js";
 import type { SessionPresence } from "./session-presence.js";
 
 const PENDING_BUFFER_MAX_BYTES = 1 * 1024 * 1024;
