+++ packages/api-server/src/apps/api-server/agent-proxies/trpc-proxy.ts
@@ -1,0 +1,106 @@
+import type { Context } from "hono";
+import type { UserIdentity } from "api-server-api";
+import { getLogger } from "../../../core/logger.js";
+import { securityLog } from "../../../core/security-log.js";
+import {
+  isAgentStoppedError,
+  isAgentWakeTimeoutError,
+} from "../../../modules/agents/index.js";
+import { podBaseUrl } from "../../../modules/agents/infrastructure/k8s.js";
+import { clientIp, hasAgentBinding, hasScope } from "../admission/auth.js";
+
+export interface AgentTrpcProxyDeps {
+  namespace: string;
+  verifyOwner: (agentId: string, ownerSub: string) => Promise<boolean>;
+  ensureReady: (agentId: string) => Promise<unknown>;
+}
+
+type ProxyCtx = Context<{ Variables: { user: UserIdentity; roles: string[] } }>;
+
+/** `/api/agents/:id/trpc/*` — proxies the agent-runtime's own tRPC surface
+ *  (in-pod file operations) to the UI. Sits behind the /api middleware
+ *  chain; adds ownership + operate-scope + binding before forwarding. */
+export function createAgentTrpcProxy(deps: AgentTrpcProxyDeps) {
+  return async (c: ProxyCtx) => {
+    const user = c.get("user");
+    const agentId = c.req.param("id")!;
+    if (!(await deps.verifyOwner(agentId, user.sub))) {
+      // The 404 is otherwise indistinguishable from a genuinely missing
+      // agent — log the cross-tenant access attempt.
+      securityLog("warn", "authz.owner_mismatch", {
+        category: "authz",
+        actor: user.sub,
+        actorKind: "user",
+        agentId,
+        decision: "deny",
+        reason: "not-owner",
+        sourceIp: clientIp(c),
+        detail: { surface: "trpc-proxy" },
+      });
+      return c.json({ error: "not found" }, 404);
+    }
+    // The in-pod relay is the most powerful surface in the system (ACP
+    // frames, pod-files, terminal). Require `agents:operate` + per-key agent
+    // binding before forwarding to the agent-runtime.
+    if (!hasScope(user, "agents:operate")) {
+      return c.json(
+        { error: "forbidden", message: "Requires agents:operate" },
+        403,
+      );
+    }
+    if (!hasAgentBinding(user, agentId)) {
+      return c.json(
+        {
+          error: "forbidden",
+          message: `API key is not bound to agent ${agentId}`,
+        },
+        403,
+      );
+    }
+
+    try {
+      await deps.ensureReady(agentId);
+    } catch (err) {
+      getLogger().warn(
+        { agentId, error: (err as Error).message },
+        "trpc-proxy.ensure-ready.failed",
+      );
+      return c.json(
+        {
+          error: "agent unreachable",
+          ...(isAgentWakeTimeoutError(err) ? { reason: err.failure.kind } : {}),
+          ...(isAgentStoppedError(err) ? { reason: "stopped" } : {}),
+        },
+        502,
+      );
+    }
+
+    // No Bearer swap needed: ownership is verified above, and the agent
+    // pod's NetworkPolicy admits ingress only from the api-server pod —
+    // the kernel-level gate is the auth boundary on this hop.
+    const rest = c.req.path.replace(`/api/agents/${agentId}/trpc`, "");
+    const qs = c.req.url.includes("?") ? "?" + c.req.url.split("?")[1] : "";
+    const upstreamUrl = `http://${podBaseUrl(agentId, deps.namespace)}/api/trpc${rest}${qs}`;
+    try {
+      const headers = new Headers(c.req.raw.headers);
+      headers.delete("host");
+      headers.delete("authorization");
+      const upstream = await fetch(upstreamUrl, {
+        method: c.req.method,
+        headers,
+        body:
+          c.req.method !== "GET" && c.req.method !== "HEAD"
+            ? c.req.raw.body
+            : undefined,
+        // @ts-expect-error -- node fetch supports duplex
+        duplex: "half",
+      });
+      return new Response(upstream.body, {
+        status: upstream.status,
+        headers: upstream.headers,
+      });
+    } catch {
+      return c.json({ error: "agent unreachable" }, 502);
+    }
+  };
+}
