+++ packages/api-server/src/apps/api-server/routes/default-brand-icon.ts
