+++ packages/api-server/src/index.ts
@@ -1,989 +1,45 @@
-import { readFileSync } from "node:fs";
-import { createDb, runMigrations } from "db";
-import { createApi } from "./modules/agents/infrastructure/k8s.js";
-import {
-  AGENTS_PLURAL,
-  LABEL_OWNER,
-} from "./modules/agents/infrastructure/labels.js";
-import {
-  composeAgentsModule,
-  createAgentsRepository,
-  createAgentEnvRepository,
-  createAgentRegistrySecretPort,
-  backfillUserEnv,
-  startChannelCleanupSaga,
-  deleteChannelsByAgent,
-  listChannelsByOwner,
-  findBySlackChannelId,
-  findSlackChannelsByAgent,
-  deleteSlackChannelBinding,
-  setSlackChannelAmbient,
-  createAgentSweep,
-} from "./modules/agents/index.js";
-import {
-  composePrStateResolver,
-  createAgentSkillsRepository,
-  parseSeedSources,
-  startSkillsCleanupSaga,
-} from "./modules/skills/index.js";
-import { createK8sClient } from "./modules/agents/infrastructure/k8s.js";
-import { createAcpClient, type AcpClientFactory } from "./core/acp-client.js";
-import { createPostgresState } from "@chat-adapter/state-pg";
-import {
-  createSlackWorker,
-  type SlackOAuthPending,
-  type ChannelRegistry,
-} from "./modules/channels/infrastructure/slack.js";
-import { createBoltSlackGateway } from "./modules/channels/infrastructure/bolt-slack-gateway.js";
-import { createFakeSlackGateway } from "./modules/channels/infrastructure/fake-slack-gateway.js";
-import { createTelegramWorker } from "./modules/channels/infrastructure/telegram.js";
-import {
-  createChannelManager,
-  type ChannelRpcRequest,
-} from "./modules/channels/services/channel-manager.js";
-import { createIdentityLinkService } from "./modules/channels/services/identity-link-service.js";
-import {
-  findIdentityByExternalUser,
-  upsertIdentityLink,
-  deleteIdentityLink,
-} from "./modules/channels/infrastructure/identity-links-repository.js";
-import {
-  findAgentByConversation,
-  bindConversation,
-  listConversationsByAgent,
-  unbindConversation,
-  deleteConversationsByAgent,
-} from "./modules/channels/infrastructure/telegram-conversations-repository.js";
-import {
-  createTelegramBindFlowStore,
-  type TelegramOAuthPending,
-} from "./modules/channels/infrastructure/telegram-flows.js";
-import { createSlackBindFlowStore } from "./modules/channels/infrastructure/slack-flows.js";
-import {
-  composeRuntimeDelivery,
-  createBullConnection,
-} from "./modules/runtime-delivery/index.js";
-import { createHarnessConfigSnapshotRepo } from "./modules/harness-config/index.js";
-import { composeSchedulesAtBoot } from "./modules/schedules/index.js";
-import {
-  createKubernetesSecretStore,
-  createSecretStoreRegistry,
-} from "./modules/secret-store/index.js";
-import { composeUsageModule } from "./modules/usage/compose.js";
-import { listAgentIdsByOwner } from "./modules/usage/infrastructure/agents-postgres-repository.js";
-import { composeMetricsReader } from "./modules/metrics/index.js";
-import { composeAuditModule } from "./modules/audit/index.js";
-import { composeE2eModule } from "./modules/e2e/compose.js";
-import { composeTermsModule } from "./modules/terms/index.js";
-import { loadConfig } from "./config.js";
-import { configureLogger, getLogger } from "./core/logger.js";
+import { bootstrap } from "./bootstrap.js";
 import { startApiServerApp } from "./apps/api-server/app.js";
 import { startHarnessApiServerApp } from "./apps/harness-api-server/app.js";
-import { composeArtifactsModule } from "./modules/artifacts/compose.js";
-import { createTemplatesRepository } from "./modules/templates/infrastructure/templates-repository.js";
-import { composeTemplatesModule } from "./modules/templates/compose.js";
-import {
-  composeInvocationLivenessSweep,
-  createDriverResolutionAdapter,
-  createInvocationsCleanupHook,
-  listInvocationAgentIds,
-} from "./modules/invocations/index.js";
 import { startExtAuthzGrpcApp } from "./apps/ext-authz/grpc.js";
-import {
-  composeApprovalsSystem,
-  createApprovalsCleanupHook,
-  listPendingApprovalAgentIds,
-} from "./modules/approvals/compose.js";
-import { createWrapperFrameSender } from "./modules/approvals/infrastructure/wrapper-frame-sender.js";
-import {
-  createEgressRuleMatchAdapter,
-  createEgressRulesCleanupHook,
-  createL7PromotionBackfill,
-  createPresetSeederAdapter,
-  listEgressRuleAgentIds,
-} from "./modules/egress-rules/compose.js";
-import {
-  composeConnectionsAtBoot,
-  composeConnectionsForOwner,
-  createConnectionGrantsCleanupHook,
-  listConnectionGrantAgentIds,
-} from "./modules/connections/compose.js";
-import { createConnectionRulesSyncAdapter } from "./modules/egress-rules/compose.js";
-import { createAgentArtifactsSweeper } from "./sagas/agent-artifacts-sweeper.js";
-import {
-  composeExperimentInactivitySweep,
-  reconcileExperimentPins,
-} from "./modules/experiments/index.js";
-import { EXPERIMENT_ACTIVE_KEY } from "./modules/agents/infrastructure/labels.js";
-import {
-  composeArtifactExpirySweeper,
-  composeArtifactLibraryForOwner,
-} from "./modules/artifact-library/index.js";
-import { createK8sClient as createAgentsK8sClient } from "./modules/agents/infrastructure/k8s.js";
-import { sweepLegacyCredentialSecrets } from "./modules/connections/sweep/legacy-secret-sweep.js";
-import { loadTrustedHosts } from "./bootstrap/trusted-hosts.js";
-import { createPeriodicJobs } from "./core/periodic-jobs.js";
-import { createRedisTtlStore } from "./core/ttl-store.js";
-import { createRedisBus } from "./core/redis-bus.js";
-import { createBusRpc } from "./core/bus-rpc.js";
-import { createRedisBlobHandoff } from "./core/blob-handoff.js";
-import { createLeaderLease } from "./core/leader-lease.js";
-import { createTurnAttendance } from "./core/turn-attendance.js";
-import { createSubPseudonymizer } from "./core/sub-pseudonymizer.js";
-import { podBaseUrl } from "./modules/agents/infrastructure/k8s.js";
 
-const config = loadConfig();
-configureLogger({
-  level: config.logLevel,
-  base: { appVersion: config.appVersion },
-});
-getLogger().info("api-server starting");
+// Process entry point and conductor: assemble everything (bootstrap), spawn
+// the three network servers, and wire teardown to process signals. All
+// construction + background-service cleanup lives in bootstrap; this file
+// owns only the servers it spawns and the shutdown ordering.
+const { apiServerDeps, harnessDeps, extAuthzDeps, cleanup } = await bootstrap();
 
-const { api } = createApi(config.namespace);
-const dbTls = {
-  ca: config.databaseCaCertPath
-    ? readFileSync(config.databaseCaCertPath, "utf8")
-    : undefined,
-};
-await runMigrations(config.databaseUrl, config.migrationsPath, dbTls);
-const { db, sql } = createDb(config.databaseUrl, dbTls);
+const { server: apiServer, trpcWs } = startApiServerApp(apiServerDeps);
+const { server: harnessApiServer } = startHarnessApiServerApp(harnessDeps);
+const { server: extAuthzGrpcServer } = await startExtAuthzGrpcApp(extAuthzDeps);
 
-// Candidate-artifact storage, shared by both app servers. ensureReady
-// provisions the bucket and fails boot fast on an unreachable store.
-const artifactsModule = composeArtifactsModule({
-  maxBytes: config.maxArtifactBytes,
-  objectStorage: config.objectStorageEndpoint
-    ? {
-        endpoint: config.objectStorageEndpoint,
-        agentEndpoint:
-          config.objectStorageAgentEndpoint ?? config.objectStorageEndpoint,
-        publicEndpoint: config.objectStoragePublicEndpoint ?? null,
-        region: config.objectStorageRegion,
-        bucket: config.objectStorageBucket,
-        forcePathStyle: config.objectStorageForcePathStyle,
-        credentials:
-          config.objectStorageAccessKeyId != null &&
-          config.objectStorageSecretAccessKey != null
-            ? {
-                accessKeyId: config.objectStorageAccessKeyId,
-                secretAccessKey: config.objectStorageSecretAccessKey,
-              }
-            : null,
-      }
-    : null,
-});
-await artifactsModule.ensureReady();
-const artifacts = artifactsModule.service;
-
-if (!config.redisUrl)
-  throw new Error("REDIS_URL is required (Redis is a platform primitive)");
-const bullConnection = createBullConnection(
-  config.redisUrl,
-  config.redisPassword ?? undefined,
-);
-const redisBus = createRedisBus(config.redisUrl, {
-  password: config.redisPassword ?? undefined,
-});
-// Dedicated general-purpose Redis client for cross-replica shared state
-// (session presence keys, OAuth/bind handoff TTL stores).
-const sharedRedis = createBullConnection(
-  config.redisUrl,
-  config.redisPassword ?? undefined,
-) as import("ioredis").Redis;
-
-// Who can answer an egress approval for an agent: the channel workers mark a
-// turn open for its duration, and the ext_authz gate reads that plus the
-// relays' session-presence keys to decide whether holding is worth anything.
-const turnAttendance = createTurnAttendance(sharedRedis);
-
-// Periodic background work runs as BullMQ job schedulers, one queue per job
-// ("periodic.<name>") — one execution per period across replicas, and a
-// hung tick can only stall its own lane. Ticks stay idempotent; the queue
-// is scheduling and visibility, never correctness. Jobs register below;
-// periodicJobs.start() runs after the last registration.
-const periodicJobs = createPeriodicJobs({
-  connection: bullConnection,
-  log: (msg) => process.stderr.write(`[periodic-jobs] ${msg}\n`),
-});
-
-const k8sClient = createK8sClient(api, config.namespace);
-const agentsRepo = createAgentsRepository(k8sClient);
-const agentEnvRepo = createAgentEnvRepository(db);
-
-// Migrate pre-existing spec.env onto agent_env. First pass runs before serving;
-// per-agent failures are isolated and retried on a timer until a clean pass.
-let backfillRetry: ReturnType<typeof setTimeout> | undefined;
-async function runUserEnvBackfill(): Promise<void> {
-  const { failed } = await backfillUserEnv({
-    repo: agentsRepo,
-    agentEnvRepo,
-    log: (m) => getLogger().info(`[user-env] ${m}`),
-  });
-  if (failed > 0)
-    backfillRetry = setTimeout(() => void runUserEnvBackfill(), 60_000);
-}
-await runUserEnvBackfill();
-
-// One-time sweep of the drained legacy credential Secrets the
-// secrets→connections migration left in place (#2198). Fire-and-forget:
-// nothing at runtime reads these Secrets, a failed pass retries on the
-// next boot, and the summary line doubles as the drain-gate confirmation.
-void sweepLegacyCredentialSecrets({
-  k8sClient,
-  log: (m) => getLogger().info(`[legacy-secret-sweep] ${m}`),
-  logError: (m) => getLogger().error(`[legacy-secret-sweep] ${m}`),
-}).catch((err) =>
-  getLogger().error(`[legacy-secret-sweep] pass failed: ${String(err)}`),
-);
-
-// Concrete spec.resources.limits on legacy agents (#1900) are the
-// controller's job: it materializes `legacyAgentSize` into any spec
-// missing a dimension on reconcile (fill-if-absent) — watch-driven, so it
-// also covers CRs created out-of-band, which a boot backfill here never
-// could.
-
-const runtimeDelivery = composeRuntimeDelivery({
-  db,
-  namespace: config.namespace,
-  bullConnection,
-  // The apply worker only dispatches to a ready agent (the controller's CRD
-  // Ready condition); otherwise it defers and the sweep retries once it's live.
-  agentRunningPort: {
-    isRunning: (agentId) => agentsRepo.isReady(agentId),
-  },
-  snapshotWriter: createHarnessConfigSnapshotRepo(db),
-  harnessServerUrl: config.harnessServerUrl,
-});
-await periodicJobs.register("runtime-outbox-sweep", 60_000, () =>
-  runtimeDelivery.sweep.tick(),
-);
-const contributionsSettledPort = {
-  status: runtimeDelivery.contributionsStatus,
-  statusMany: runtimeDelivery.contributionsStatusMany,
-  isSettled: async (agentId: string) =>
-    (await runtimeDelivery.contributionsStatus(agentId)).settled,
-};
-const subPseudonymizer = createSubPseudonymizer(config.activityHmacKey);
-
-const secretStores = createSecretStoreRegistry();
-secretStores.register(createKubernetesSecretStore({ k8s: k8sClient }));
-
-// Redis-backed handoff stores: the browser callback leg may land on any replica.
-const OAUTH_FLOW_TTL_MS = 10 * 60 * 1000;
-const connectionsBoot = composeConnectionsAtBoot({
-  db,
-  secretStore: secretStores.default(),
-  // Same Redis namespace as the app-side engine — flows are one namespace
-  // regardless of which composition minted them.
-  pendingFlowStore: createRedisTtlStore(
-    sharedRedis,
-    "oauth:connections",
-    OAUTH_FLOW_TTL_MS,
-  ),
-});
-const connectionsServiceFor = (ownerId: string) =>
-  composeConnectionsForOwner({
-    ownerId,
-    db,
-    templates: connectionsBoot.templates,
-    oauthEngine: connectionsBoot.oauthEngine,
-    githubAppEngine: connectionsBoot.githubAppEngine,
-    secretStore: secretStores.default(),
-    runtimeMutator: runtimeDelivery.runtimeMutator,
-    agentsRepo,
-    connectionRulesSync: createConnectionRulesSyncAdapter(db),
-    oauthCallbackUrl: `${config.uiBaseUrl}/api/oauth/callback`,
-    brandName: config.brand.name,
-  });
-// L7-promotion backfill (#2865): project active narrow rules onto each
-// Agent CR's spec.l7Hosts and retire the legacy owner-scoped allow-only
-// marker Secrets. Same capped-retry shape as the secrets migration; while
-// it hasn't fully succeeded the markers stay and rules remain enforced.
-let l7BackfillRetry: ReturnType<typeof setTimeout> | undefined;
-let l7BackfillRetries = 0;
-const L7_BACKFILL_MAX_RETRIES = 10;
-const l7PromotionBackfill = createL7PromotionBackfill(db, k8sClient, (m) =>
-  getLogger().info(`[l7-backfill] ${m}`),
-);
-async function runL7PromotionBackfill(): Promise<void> {
-  let failed: number;
-  try {
-    ({ failed } = await l7PromotionBackfill());
-  } catch (err) {
-    // A pass can throw outside its per-agent handling (e.g. listing the
-    // legacy markers while K8s is flaky) — treat it as a failed pass so
-    // the timer retry never becomes an unhandled rejection.
-    getLogger().error(`[l7-backfill] pass failed: ${String(err)}`);
-    failed = 1;
-  }
-  if (failed === 0) return;
-  if (l7BackfillRetries >= L7_BACKFILL_MAX_RETRIES) {
-    getLogger().error(
-      `[l7-backfill] still ${failed} failing after ` +
-        `${L7_BACKFILL_MAX_RETRIES} retries; giving up — legacy allow-only ` +
-        `markers stay in place and rules remain enforced`,
-    );
-    return;
-  }
-  l7BackfillRetries++;
-  l7BackfillRetry = setTimeout(() => void runL7PromotionBackfill(), 60_000);
-}
-await runL7PromotionBackfill();
-
-const { service: termsService, isAcceptedPort: isTermsAccepted } =
-  composeTermsModule({
-    db,
-    version: config.terms.version,
-    text: config.terms.text,
-  });
-
-const fakeSlackGateway =
-  config.e2eEnabled && !(config.slackBotToken && config.slackAppToken)
-    ? createFakeSlackGateway()
-    : undefined;
-
-const { service: e2eService } = composeE2eModule({
-  namespace: config.namespace,
-  slack: fakeSlackGateway,
-});
-
-const channelCleanupSub = startChannelCleanupSaga(
-  deleteChannelsByAgent(db),
-  deleteConversationsByAgent(db),
-);
-const skillsCleanupSub = startSkillsCleanupSaga((agentId) =>
-  createAgentSkillsRepository(db).deleteByAgent(agentId),
-);
-const seedSources = parseSeedSources(config.skillSourcesSeed);
-
-const usage = composeUsageModule({
-  db,
-  subPseudonymizer,
-  activityTrackingEnabled: config.activityTrackingEnabled,
-  inspectorRole: config.keycloakInspectorRole ?? "",
-  listK8sAgents: async () => {
-    const agents = await k8sClient.listCustomObjects(AGENTS_PLURAL);
-    return agents
-      .filter((a) => a.metadata?.name && a.metadata?.labels?.[LABEL_OWNER])
-      .map((a) => ({
-        id: a.metadata!.name!,
-        owner: a.metadata!.labels![LABEL_OWNER]!,
-      }));
-  },
-});
-usage.start();
-
-// Security audit trail (bus-driven half). Denials and call-site-only
-// mutations log directly at their sites; this covers the actor-bearing
-// success/observation events on the domain bus.
-const audit = composeAuditModule();
-audit.start();
-
-const { agents: systemAgents } = composeAgentsModule({
-  api,
-  namespace: config.namespace,
-  agentIdleTimeoutMinutes: config.agentIdleTimeoutMinutes,
-  agentDefaultLimits: {
-    cpu: config.agentDefaultCpuLimit,
-    memory: config.agentDefaultMemoryLimit,
-  },
-  owner: undefined,
-  db,
-  readTemplateSpec: async () => null,
-  runtimeMutator: runtimeDelivery.runtimeMutator,
-  contributionsSettled: contributionsSettledPort,
-});
-
-const identityLinkService = createIdentityLinkService({
-  findByExternalUser: findIdentityByExternalUser(db),
-  upsert: upsertIdentityLink(db),
-  delete: deleteIdentityLink(db),
-});
-
-const pendingSlackOAuthFlows = createRedisTtlStore<SlackOAuthPending>(
-  sharedRedis,
-  "oauth:slack",
-  OAUTH_FLOW_TTL_MS,
-);
-const pendingTelegramOAuthFlows = createRedisTtlStore<TelegramOAuthPending>(
-  sharedRedis,
-  "oauth:telegram",
-  OAUTH_FLOW_TTL_MS,
-);
-const telegramBindFlows = config.telegramBotToken
-  ? createTelegramBindFlowStore({
-      store: createRedisTtlStore(
-        sharedRedis,
-        "bind:telegram",
-        OAUTH_FLOW_TTL_MS,
-      ),
-    })
-  : undefined;
-// Unconditional: needed whenever the Slack OAuth callback is mounted (real
-// tokens or e2e). The bind command mints entries.
-const slackBindFlows = createSlackBindFlowStore({
-  store: createRedisTtlStore(sharedRedis, "bind:slack", OAUTH_FLOW_TTL_MS),
-});
-const slackOauthCallbackUrl =
-  config.slackOauthCallbackUrl ??
-  `${config.uiBaseUrl}/api/slack/oauth/callback`;
-const telegramOauthCallbackUrl = `${config.uiBaseUrl}/api/telegram/oauth/callback`;
-
-// The chat-sdk state pool uses node-postgres (`pg`), which — unlike postgres-js
-// — reads a CA file from `sslrootcert` in the connection string. Append it so
-// the pool verifies against the same scoped CA; trust stays on this connection.
-const chatSdkDatabaseUrl = config.databaseCaCertPath
-  ? `${config.databaseUrl}${config.databaseUrl.includes("?") ? "&" : "?"}sslrootcert=${config.databaseCaCertPath}`
-  : config.databaseUrl;
-const chatSdkState = config.telegramBotToken
-  ? createPostgresState({ url: chatSdkDatabaseUrl, keyPrefix: "chat-sdk" })
-  : undefined;
-
-const channelRegistry: ChannelRegistry = {
-  resolveSlackBinding: async (slackChannelId) => {
-    const row = await findBySlackChannelId(db)(slackChannelId);
-    if (!row) return null;
-    return {
-      instanceName: row.agentId,
-      owner: row.owner,
-      ...(row.ambient ? { ambient: true } : {}),
-    };
-  },
-  resolveSlackChannelsByInstance: findSlackChannelsByAgent(db),
-};
-
-const slackTokens =
-  config.slackBotToken && config.slackAppToken
-    ? { botToken: config.slackBotToken, appToken: config.slackAppToken }
-    : null;
-
-const slackGatewayFactory = slackTokens
-  ? () =>
-      createBoltSlackGateway({
-        botToken: slackTokens.botToken,
-        appToken: slackTokens.appToken,
-        commandName: `/${config.brand.short}`,
-      })
-  : fakeSlackGateway
-    ? () => fakeSlackGateway
-    : undefined;
-
-// Bind the ACP turn ceiling (and namespace) into the client factories at the
-// composition root; workers get pre-wired factories and stay ignorant of both.
-const acpTurnCeilingMs = config.acpTurnCeilingSeconds * 1000;
-const makeAcpClient: AcpClientFactory = (instanceName) =>
-  createAcpClient({
-    namespace: config.namespace,
-    instanceName,
-    turnCeilingMs: acpTurnCeilingMs,
-  });
-
-const slackWorker = slackGatewayFactory
-  ? createSlackWorker(
-      makeAcpClient,
-      slackGatewayFactory,
-      () => systemAgents,
-      identityLinkService,
-      {
-        keycloakExternalUrl: config.keycloakExternalUrl,
-        keycloakUrl: config.keycloakUrl,
-        keycloakRealm: config.keycloakRealm,
-        keycloakClientId: config.keycloakClientId,
-        callbackUrl: slackOauthCallbackUrl,
-      },
-      pendingSlackOAuthFlows,
-      (agentId) => agentsRepo.getOwner(agentId),
-      channelRegistry,
-      deleteSlackChannelBinding(db),
-      setSlackChannelAmbient(db),
-      { name: config.brand.name, short: config.brand.short },
-      isTermsAccepted,
-      config.uiBaseUrl,
-      turnAttendance,
-    )
-  : undefined;
-
-const telegramWorker =
-  config.telegramBotToken && chatSdkState
-    ? createTelegramWorker({
-        botToken: config.telegramBotToken,
-        configuredBotUsername: config.telegramBotUsername,
-        makeAcpClient,
-        state: chatSdkState,
-        agents: () => systemAgents,
-        conversations: {
-          findAgentByConversation: findAgentByConversation(db),
-          bind: bindConversation(db),
-          listByAgent: listConversationsByAgent(db),
-          unbind: unbindConversation(db),
-        },
-        oauthConfig: {
-          keycloakExternalUrl: config.keycloakExternalUrl,
-          keycloakUrl: config.keycloakUrl,
-          keycloakRealm: config.keycloakRealm,
-          keycloakClientId: config.keycloakClientId,
-          callbackUrl: telegramOauthCallbackUrl,
-        },
-        pendingOAuthFlows: pendingTelegramOAuthFlows,
-        isTermsAccepted,
-        uiBaseUrl: config.uiBaseUrl,
-        brandShort: config.brand.short,
-        brandName: config.brand.name,
-        attendance: turnAttendance,
-      })
-    : undefined;
-
-// Channel workers are single-holder: Slack Socket Mode load-balances each
-// event across the app's open connections and Telegram admits one getUpdates
-// consumer per token, while every piece of per-turn state in both workers
-// (in-flight/lingering turn refs, the ambient queues, the per-thread turn
-// lock) is in-process. So one replica runs them, elected by lease, and the
-// outbound half — an agent's reply/react, which lands wherever the harness
-// Service pinned its gateway — hops to that replica over the bus.
-const channelRpc = createBusRpc<ChannelRpcRequest, unknown>({
-  bus: redisBus,
-  service: "channels",
-  // Exactly-once execution across a lease handover, when the outgoing and
-  // incoming holders both briefly serve. Without it both would run the
-  // request and post the message twice.
-  claim: async (id) =>
-    (await sharedRedis.set(`rpc:claim:channels:${id}`, "1", "EX", 60, "NX")) ===
-    "OK",
-});
-
-const channelManager = createChannelManager({
-  slackWorker,
-  telegramWorker,
-  rpc: channelRpc,
-  blobs: createRedisBlobHandoff(sharedRedis),
-  // Read per dispatch, so it always reflects the live lease rather than
-  // whatever was true at construction.
-  isLeader: () => channelLease.isLeader(),
-});
-
-const channelLease = createLeaderLease({
-  redis: sharedRedis,
-  name: "channels",
-  onAcquired: async () => {
-    // Re-read the bindings on every acquisition: this replica may be taking
-    // over minutes after boot, from a leader that died.
-    const channelsByInstance = await listChannelsByOwner(db, "")();
-    await channelManager.bootstrap(channelsByInstance);
-  },
-  onLost: () => channelManager.standDown(),
-});
-
-// Seed list for the `trusted` egress preset.
-// Read once at boot; the helm ConfigMap is the operator-editable source.
-const trustedHosts = loadTrustedHosts(config.trustedHostsPath);
-const presetSeeder = createPresetSeederAdapter(db, trustedHosts);
-
-// Egress Aliasing resolution for the ext_authz gate: Invocation target →
-// root driver, consulted before every identity resolve on the egress path.
-const invocationDriverResolution = createDriverResolutionAdapter(db);
-
-const wrapperFrameSender = createWrapperFrameSender({
-  resolveWrapperUrl: (agentId) =>
-    `ws://${podBaseUrl(agentId, config.namespace)}/api/acp`,
-});
-
-// System-level approvals composition — bound to the bus + cross-module
-// ports for instance identity (agents), rule matching (egress-rules), and
-// wrapper-frame delivery. Relay, gate, and sweeper are long-lived and
-// shared across all owners.
-const {
-  relay: approvalsRelay,
-  gate: extAuthzGate,
-  sweeper: deliverySweeper,
-} = composeApprovalsSystem({
-  db,
-  bus: redisBus,
-  identityResolver: {
-    // Egress Aliasing: an Invocation target has no egress identity of its
-    // own — resolve the caller to its root driver first, so rule match,
-    // HITL hold, and approval writes all land on the driver. A target whose
-    // driver is gone resolves to nothing and the gate fails closed.
-    resolve: async (agentId) => {
-      const rootId = await invocationDriverResolution.resolveRoot(agentId);
-      if (!rootId) return null;
-      const r = await agentsRepo.resolveIdentity(rootId);
-      return r ? { ownerSub: r.owner, agentId: r.agentId } : null;
-    },
-  },
-  ruleMatcher: {
-    match: async (agentId, host, method, path) => {
-      const matched = await createEgressRuleMatchAdapter(db).match(
-        agentId,
-        host,
-        method,
-        path,
-      );
-      return matched ? { verdict: matched.verdict } : null;
-    },
-  },
-  attendance: turnAttendance,
-  wrapperFrameSender,
-  holdSeconds: config.approvalHoldSeconds,
-  // The presigned link is the per-request authorization for the store, so
-  // no HITL decision. Bare hostname — the gate strips ports.
-  platformAllowedHosts: config.objectStorageAgentEndpoint
-    ? [new URL(config.objectStorageAgentEndpoint).hostname]
-    : [],
-});
-await periodicJobs.register("approvals-delivery-sweep", 30_000, () =>
-  deliverySweeper.tick(),
-);
-
-// Per-agent cleanup hooks fired after a successful K8s delete. Each
-// module's adapter clears its own table; failures log + continue. The
-// orphan-sweeper saga catches anything missed (replica died mid-delete,
-// hook threw, etc.).
-const agentsCleanupK8s = createAgentsK8sClient(api, config.namespace);
-const registrySecretPort = createAgentRegistrySecretPort(agentsCleanupK8s);
-const connectionGrantsCleanupHook = createConnectionGrantsCleanupHook(db);
-
-const agentEnvCleanupHook = (agentId: string) =>
-  agentEnvRepo.deleteForAgent(agentId);
-
-// Driver Cascade: deleting an agent fails its running Invocations and reaps
-// their targets, so no dangling target keeps running against a deleted
-// driver's egress identity. `agentsFor` binds lazily — the factory below is
-// declared later and composes these same hooks, which is what makes chained
-// Invocations unwind transitively.
-const invocationsCleanupHook = createInvocationsCleanupHook({
-  db,
-  agentsFor: (owner) => harnessAgentsServiceFor(owner),
-});
-
-const agentCleanupHooks = [
-  createEgressRulesCleanupHook(db),
-  createApprovalsCleanupHook(db),
-  (agentId: string) => registrySecretPort.delete(agentId),
-  connectionGrantsCleanupHook,
-  agentEnvCleanupHook,
-  invocationsCleanupHook,
-];
-
-// Cross-store orphan reaper. Lists live agent CRs, finds DB rows keyed by
-// an agent_id no longer in the live set, and runs each module's cleanup.
-// Scheduled on the periodic-jobs queue below.
-const agentArtifactsSweeper = createAgentArtifactsSweeper({
-  k8s: agentsCleanupK8s,
-  sources: [
-    {
-      name: "egress-rules",
-      listAgentIds: () => listEgressRuleAgentIds(db),
-      cleanup: agentCleanupHooks[0]!,
-    },
-    {
-      name: "pending-approvals",
-      listAgentIds: () => listPendingApprovalAgentIds(db),
-      cleanup: agentCleanupHooks[1]!,
-    },
-    {
-      name: "registry-pull-secrets",
-      listAgentIds: () => registrySecretPort.listAgentIds(),
-      cleanup: agentCleanupHooks[2]!,
-    },
-    {
-      name: "connection-grants",
-      listAgentIds: () => listConnectionGrantAgentIds(db),
-      cleanup: connectionGrantsCleanupHook,
-    },
-    {
-      name: "agent-env",
-      listAgentIds: () => agentEnvRepo.listAgentIds(),
-      cleanup: agentEnvCleanupHook,
-    },
-    {
-      name: "invocations",
-      listAgentIds: () => listInvocationAgentIds(db),
-      cleanup: invocationsCleanupHook,
-    },
-  ],
-  batchSize: 200,
-});
-
-// Experiments v2 inactivity sweep: reaps `running` Experiments whose script
-// has gone silent (no trace event) past the configured window to `failed`, so
-// every executed Experiment reaches a terminal state — and releases the
-// driver's hibernation pin when it was its last running experiment. Atomic
-// conditional transitions make it multi-replica safe. Sweep at the window
-// cadence, capped at 5 minutes.
-const experimentPin = {
-  set: (agentId: string) =>
-    agentsRepo.patchAnnotation(agentId, EXPERIMENT_ACTIVE_KEY, "true"),
-  clear: (agentId: string) =>
-    agentsRepo.patchAnnotation(agentId, EXPERIMENT_ACTIVE_KEY, ""),
-};
-const experimentInactivityMs = config.experimentInactivitySeconds * 1000;
-const experimentInactivitySweep = composeExperimentInactivitySweep({
-  db,
-  inactivityMs: experimentInactivityMs,
-  batchSize: 200,
-  pin: experimentPin,
-  artifactLibraryFor: (owner) =>
-    composeArtifactLibraryForOwner({
-      db,
-      artifacts,
-      owner,
-      shareBaseUrl: config.shareBaseUrl,
-    }).artifactLibrary,
-});
-await periodicJobs.register(
-  "experiment-inactivity-sweep",
-  Math.min(experimentInactivityMs, 5 * 60_000),
-  () => experimentInactivitySweep.tick(),
-);
-
-// Converge experiment pins to database truth — experiments survive restarts
-// (unlike sessions), so this reconciles rather than blanket-clears.
-void reconcileExperimentPins({
-  db,
-  listPinnedAgentIds: () =>
-    agentsRepo.listAgentIdsWithAnnotation(EXPERIMENT_ACTIVE_KEY, "true"),
-  pin: experimentPin,
-}).then(
-  ({ set, cleared }) => {
-    if (set > 0 || cleared > 0) {
-      process.stderr.write(
-        `[experiments] pin reconciliation: set ${set}, cleared ${cleared}\n`,
-      );
+let shuttingDown = false;
+const onSignal = (): void => {
+  if (shuttingDown) return;
+  shuttingDown = true;
+  void (async () => {
+    process.stderr.write("shutting down...\n");
+    // Nudge events-socket clients onto surviving replicas first, then drop
+    // the sockets this replica holds.
+    trpcWs.drain();
+    await trpcWs.close();
+    // Background services and stores (bootstrap-owned).
+    await cleanup();
+    // Then the servers this file spawned.
+    extAuthzGrpcServer.tryShutdown(() => {});
+    harnessApiServer.close();
+    apiServer.close();
+    // Flush OTel if the --import bootstrap (dist/telemetry.js) registered it.
+    // Reached via Symbol lookup — importing telemetry.ts here would evaluate a
+    // second copy of that bundle and split the SDK singleton.
+    const flushOtel = (globalThis as Record<symbol, unknown>)[
+      Symbol.for("platform.otel.shutdown")
+    ];
+    if (typeof flushOtel === "function") {
+      await (flushOtel as () => Promise<void>)();
     }
-  },
-  (err) => {
-    process.stderr.write(`[experiments] pin reconciliation failed: ${err}\n`);
-  },
-);
-
-// Cross-store orphan reap every 30 minutes — cheap diff scans, orphans are
-// rare and non-urgent.
-await periodicJobs.register("agent-artifacts-sweep", 30 * 60_000, () =>
-  agentArtifactsSweeper.tick(),
-);
-
-// Artifact-library expiry sweep: hard-deletes artifacts (private ones too)
-// whose expiry passed more than the grace window ago (the viewer 410s public
-// ones meanwhile). Hourly is plenty — expiry granularity is hours.
-const artifactExpirySweeper = composeArtifactExpirySweeper({
-  db,
-  artifacts: artifactsModule.service,
-  batchSize: 200,
-});
-await periodicJobs.register("artifact-expiry-sweep", 60 * 60_000, () =>
-  artifactExpirySweeper.tick(),
-);
-
-// Re-read the pull requests behind skill publish records so the badge reports
-// their real state. Ten minutes is the tick, but each record is re-checked at
-// most hourly (the resolver owns that throttle) — the short tick exists so a
-// freshly published pull request resolves promptly rather than waiting an hour.
-const prStateResolver = composePrStateResolver({
-  db,
-  agents: agentsRepo,
-  namespace: config.namespace,
-  log: (msg) => process.stderr.write(`[pr-state-resolver] ${msg}\n`),
-});
-await periodicJobs.register("skill-pr-state-resolve", 10 * 60_000, () =>
-  prStateResolver.tick(),
-);
-periodicJobs.start();
-
-const schedulesBoot = composeSchedulesAtBoot({
-  db,
-  bullConnection,
-  runtimeMutator: runtimeDelivery.runtimeMutator,
-  wakeAgent: async (agentId) => {
-    await agentsRepo.wakeIfHibernated(agentId);
-  },
-});
-schedulesBoot.runner.restoreAll().catch((err) => {
-  process.stderr.write(
-    `[schedules] restoreAll failed: ${(err as Error).message}\n`,
-  );
-});
-
-// Stale active-session pins (a replica that died holding sessions) are
-// cleared by the session-presence reconcile periodic job — a boot-time
-// blanket clear would wrongly drop pins held by other live replicas.
-
-// Owner-scoped agents factory for the harness surface (the spawn primitive):
-// a driver agent spawns an Invocation target = create an ephemeral Agent +
-// submit a prompt. Mirrors the main app's per-owner agents composition, incl.
-// the connections grantProvisioner so a target's connection subset
-// materializes on create.
-const harnessTemplatesRepo = createTemplatesRepository(
-  config.agentTemplatesPath,
-);
-const { readSpec: harnessReadTemplateSpec } =
-  composeTemplatesModule(harnessTemplatesRepo);
-const wakeAgentFor = async (agentId: string) => {
-  await agentsRepo.wakeIfHibernated(agentId);
+    process.exit(0);
+  })();
 };
-const harnessAgentsServiceFor = (owner: string) => {
-  const connections = connectionsServiceFor(owner);
-  return composeAgentsModule({
-    api,
-    namespace: config.namespace,
-    agentIdleTimeoutMinutes: config.agentIdleTimeoutMinutes,
-    virtualizationEnabled: config.virtualizationEnabled,
-    agentDefaultLimits: {
-      cpu: config.agentDefaultCpuLimit,
-      memory: config.agentDefaultMemoryLimit,
-    },
-    owner,
-    db,
-    readTemplateSpec: harnessReadTemplateSpec,
-    presetSeeder,
-    cleanupHooks: agentCleanupHooks,
-    runtimeMutator: runtimeDelivery.runtimeMutator,
-    contributionsSettled: contributionsSettledPort,
-    grantProvisioner: {
-      resolveSpecGrants(sel) {
-        return Promise.resolve({
-          grantedConnectionIds: Array.from(new Set(sel.connectionIds)),
-        });
-      },
-      async applyAfterCreate(agentId, sel) {
-        if (sel.connectionIds.length)
-          await connections.setAgentConnections(agentId, sel.connectionIds);
-      },
-    },
-  }).agents;
-};
-
-// Invocation liveness sweep (owner-agnostic; started once): bounds one result
-// — fails a target that ran past its deadline or crashed mid-turn, reaps it,
-// and drops aged result rows. It does NOT own the target agent's lifecycle.
-const invocationLivenessSweep = composeInvocationLivenessSweep({
-  db,
-  agentsFor: harnessAgentsServiceFor,
-  k8s: k8sClient,
-  batchSize: 200,
-});
-await periodicJobs.register("invocation-liveness-sweep", 60_000, () =>
-  invocationLivenessSweep.tick(),
-);
-
-// Agent Sweep (#2816): generic GC that deletes any Sweepable agent once it
-// hibernates (after its Lifetime grace, if any). Keyed off Agent state, never
-// the Invocations table — the backstop for Invocation targets and the home for
-// future Sweepable agents (e.g. inherited channel agents). Owner-agnostic:
-// lists all agents, resolves an owner-scoped service per agent to delete.
-const agentSweep = createAgentSweep({
-  listAgents: () => agentsRepo.list(),
-  agentsFor: harnessAgentsServiceFor,
-});
-await periodicJobs.register("agent-sweep", 60_000, () => agentSweep.tick());
-
-const { server: apiServer } = startApiServerApp({
-  periodicJobs,
-  sharedRedis,
-  config,
-  api,
-  db,
-  channelManager,
-  identityLinkService,
-  pendingSlackOAuthFlows,
-  pendingTelegramOAuthFlows,
-  telegramBindFlows,
-  slackBindFlows,
-  seedSources,
-  redisBus,
-  approvalsRelay,
-  wrapperFrameSender,
-  presetSeeder,
-  trustedHosts,
-  agentCleanupHooks,
-  secretStores,
-  runtimeMutator: runtimeDelivery.runtimeMutator,
-  contributionsSettled: contributionsSettledPort,
-  getAgentCapabilities: (agentId) =>
-    runtimeDelivery.agentsRuntimeRepo
-      .get(agentId)
-      .then((r) => r?.runtimeCapabilities ?? null),
-  schedulesBoot,
-  mountUsageRoutes: usage.mount,
-  listRegisteredAgentIds: listAgentIdsByOwner(db, subPseudonymizer),
-  metricsReader: composeMetricsReader(config),
-  terms: termsService,
-  isTermsAccepted,
-  e2e: e2eService,
-  artifacts,
-});
-
-const { server: harnessApiServer } = startHarnessApiServerApp({
-  config,
-  api,
-  db,
-  channelManager,
-  seedSources,
-  runtimeHello: runtimeDelivery.hello,
-  schedulesBoot,
-  runtimeMutator: runtimeDelivery.runtimeMutator,
-  artifacts,
-  agentsServiceFor: harnessAgentsServiceFor,
-  connectionsServiceFor,
-  wakeAgent: wakeAgentFor,
-  runtimeSettled: contributionsSettledPort,
-});
-
-// Instance identity for ext-authz now flows from the per-instance
-// ext-authz Service the gateway pod's Envoy was configured to dial,
-// cryptographically pinned by the AuthorizationPolicy on each per-instance
-// Service. The pod-IP resolver and `x-platform-instance` header are gone.
-//
-// Single gRPC ext_authz server serves both Envoy filters: HTTP filter on
-// TLS-terminated chains (L7 — sees method/path) and the network filter on
-// the catch-all chain (L4 — SNI only). Same Check RPC, same gate service;
-// the handler reads what's populated and falls back to wildcards otherwise.
-const { server: extAuthzGrpcServer } = await startExtAuthzGrpcApp({
-  port: config.extAuthzPort,
-  holdSeconds: config.approvalHoldSeconds,
-  gate: extAuthzGate,
-  releaseName: config.releaseName,
-});
-
-// The bot handle renders in bind instructions on any replica, so resolve it
-// everywhere; only the lease holder actually runs the poller.
-void telegramWorker?.resolveIdentity();
-// Campaigns for the channel lease and bootstraps the workers if it wins.
-void channelLease.start();
-
-async function shutdown() {
-  process.stderr.write("shutting down...\n");
-  if (backfillRetry) clearTimeout(backfillRetry);
-  if (l7BackfillRetry) clearTimeout(l7BackfillRetry);
-  channelCleanupSub.unsubscribe();
-  skillsCleanupSub.unsubscribe();
-  usage.stop();
-  audit.stop();
-  await periodicJobs.close();
-  // Release the lease before tearing the workers down, so a surviving replica
-  // can pick the channels up without waiting out the TTL.
-  await channelLease.stop();
-  channelRpc.close();
-  await channelManager.stopAll();
-  await runtimeDelivery.worker.close();
-  await runtimeDelivery.queue.close();
-  await schedulesBoot.close();
-  await redisBus.close();
-  turnAttendance.close();
-  await sharedRedis.quit().catch(() => {});
-  await sql.end();
-  extAuthzGrpcServer.tryShutdown(() => {});
-  harnessApiServer.close();
-  apiServer.close();
-  // Flush OTel if the --import bootstrap (dist/telemetry.js) registered it.
-  // Reached via Symbol lookup — importing telemetry.ts here would evaluate a
-  // second copy of that bundle and split the SDK singleton.
-  const flushOtel = (globalThis as Record<symbol, unknown>)[
-    Symbol.for("platform.otel.shutdown")
-  ];
-  if (typeof flushOtel === "function") {
-    await (flushOtel as () => Promise<void>)();
-  }
-  process.exit(0);
-}
-process.on("SIGTERM", shutdown);
-process.on("SIGINT", shutdown);
+process.on("SIGTERM", onSignal);
+process.on("SIGINT", onSignal);
