+++ packages/api-server/src/modules/agents/index.ts
@@ -13,7 +13,6 @@ export {
   createAgentEnvRepository,
   type AgentEnvRepository,
 } from "./infrastructure/agent-env-repository.js";
-export { backfillUserEnv } from "./services/backfill-user-env.js";
 export {
   createAgentRegistrySecretPort,
   type AgentRegistrySecretPort,
