+++ packages/api-server/src/modules/connections/index.ts
@@ -2,3 +2,4 @@ export {
   createConnectionsRepository,
   type ConnectionsRepository,
 } from "./infrastructure/connections-repository.js";
+export { createOAuthRoutes, type OAuthCallbackDeps } from "./oauth-routes.js";
